+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # browser-switch-android Release Notes
 
+## unreleased
+
+* Add `appLinkUri` to `BrowserSwitchOptions` for Android App Link support
+
 ## 2.6.1
 
 * Throw `BrowserSwitchException` when a browser is not found to start browser switch